Fixed errors in client console.log()
