+++ backend/src/procedures/permissionProcedure.ts
@@ -3,11 +3,11 @@ import throwAndLogTRPCError from '../libs/throwAndLogTRPCError'
 import { userHasPermissions } from '../services/permission'
 import authenticatedProcedure from './authenticatedProcedure'
 
-const permissionProcedure = (permissions: Permission[]) =>
+const permissionProcedure = (permissions: Permission[], shouldLog = true) =>
   authenticatedProcedure.use(async ({ ctx, next }) => {
     const hasPermission = await userHasPermissions(ctx.user.id, permissions)
 
-    if (!hasPermission) return throwAndLogTRPCError('FORBIDDEN', 'Forbidden', `User ${ctx.user.id} tried to access a route without permissions ${permissions.join(', ')}`, 'warn')
+    if (!hasPermission) return throwAndLogTRPCError('FORBIDDEN', 'forbidden', `User ${ctx.user.username} tried to access a route without the required permissions ${permissions.join(', ')}.`, 'warn')
 
     return next({ ctx })
   })
