+++ backend/src/routers/permissionRouter.ts
@@ -1,5 +1,6 @@
 import canCreateInvitation from '../routes/permission/canCreateInvitationQuery'
 import canDeleteUser from '../routes/permission/canDeleteUserQuery'
+import canReadRoles from '../routes/permission/canReadRolesRoute'
 import canResetPassword from '../routes/permission/canResetPasswordQuery'
 import getAllPermissions from '../routes/permission/getAllPermissionsQuery'
 import { t } from '../services/trpc'
@@ -9,5 +10,6 @@ export const permissionRouter = t.router({
 
   canDeleteUser,
   canResetPassword,
-  canCreateInvitation
+  canCreateInvitation,
+  canReadRoles
 })
