+++ backend/src/routes/common/permissionCheckQuery.ts
@@ -1,0 +1,12 @@
+import { Permission } from '@prisma/client'
+import authenticatedProcedure from '../../procedures/authenticatedProcedure'
+import { userHasPermissions } from '../../services/permission'
+
+export const permissionCheckQuery = (permission: Permission) =>
+  authenticatedProcedure.query(async ({ ctx }) => {
+    const hasPermission = await userHasPermissions(ctx.user.id, [permission])
+
+    if (!hasPermission) return false
+
+    return true
+  })
