+++ backend/src/routes/permission/canReadRoles.ts
@@ -1,0 +1,4 @@
+import { permissionCheckQuery } from "../common/permissionCheckQuery";
+
+
+const canReadRolesRoute = permissionCheckQuery('ROLES_READ')
