+++ backend/src/routes/permission/canReadRolesRoute.ts
@@ -1,0 +1,5 @@
+import { permissionCheckQuery } from '../common/permissionCheckQuery'
+
+const canReadRolesRoute = permissionCheckQuery('ROLES_READ')
+
+export default canReadRolesRoute
