+++ frontend/src/routes/(application)/+page.svelte
@@ -1,72 +1,13 @@
 <script lang="ts">
-  import CopyField from '../../components/CopyField.svelte'
-  import ErrorMessage from '../../components/ErrorMessage.svelte'
-  import IconTextButton from '../../components/IconTextButton.svelte'
-  import Loader from '../../components/Loader.svelte'
-  import Modal from '../../components/Modal.svelte'
-  import RoleListItem from '../../components/RoleListItem.svelte'
-  import UserListItem from '../../components/UserListItem.svelte'
-  import { createUserInvitation } from '../../services/invitation'
-  import { canCreateInvitation } from '../../services/permission'
-  import { getAllRoles } from '../../services/role'
-  import { getUsers } from '../../services/user'
+  import RoleList from '../../components/RoleList.svelte'
+  import UserList from '../../components/UserList.svelte'
+  import { canReadRoles } from '../../services/permission'
 
-  const users = getUsers()
-  const createUserInvitationMutation = createUserInvitation()
-  const canInviteUser = canCreateInvitation()
-  const roles = getAllRoles()
-
-  let inviteLink = ''
-  let invitationOpen = false
-
-  const inviteUser = async () => {
-    const invitationToken = await $createUserInvitationMutation.mutateAsync()
-    inviteLink = `${window.location.origin}/register?invitationToken=${encodeURIComponent(invitationToken)}`
-    invitationOpen = true
-  }
+  const canReadRolesQuery = canReadRoles()
 </script>
 
-<div class="w-screen p-4 lg:w-2/3 2xl:w-1/2 lg:m-auto">
-  <div class="flex flex-row items-center justify-between mb-2">
-    <span class="text-2xl font-bold">Users</span>
+<UserList />
 
-    {#if $canInviteUser.data}
-      <IconTextButton on:click={inviteUser} icon="ic:round-plus" iconClass="w-6 h-6" title="Add User" />
-      <Modal bind:open={invitationOpen}>
-        <h4 class="text-lg font-bold">Invitation link</h4>
-        <p class="text-sm text-gray-400">The link is automatically copied to your clipboard.</p>
-        <CopyField value={inviteLink} copyOnMount />
-      </Modal>
-    {/if}
-  </div>
-
-  <div class="flex flex-col p-2 rounded-md bg-slate-900">
-    {#if $users.isLoading}
-      <Loader />
-    {:else if $users.error}
-      <ErrorMessage errorMessage="Could not load users" />
-    {:else}
-      {#each $users.data as user}
-        <UserListItem {user} />
-      {/each}
-    {/if}
-  </div>
-</div>
-
-<div class="w-screen p-4 lg:w-2/3 2xl:w-1/2 lg:m-auto">
-  <div class="flex flex-row items-center justify-between mb-2">
-    <span class="text-2xl font-bold">Roles</span>
-  </div>
-
-  <div class="flex flex-col p-2 rounded-md bg-slate-900">
-    {#if $roles.isLoading}
-      <Loader />
-    {:else if $roles.error}
-      <ErrorMessage errorMessage="Could not load roles" />
-    {:else}
-      {#each $roles.data as role}
-        <RoleListItem {role} />
-      {/each}
-    {/if}
-  </div>
-</div>
+{#if $canReadRolesQuery.data}
+  <RoleList />
+{/if}
